Initialized project.
